1383 tech debt (#2414)
* Adjust for new states

* Move helpers
