+++ cosmetics-web/spec/features/notifications_deleting_spec.rb
@@ -15,7 +15,8 @@ RSpec.describe "Notifications delete", type: :feature do
     draft_notification
     visit "/responsible_persons/#{responsible_person.id}/notifications"
     click_on draft_notification.product_name
-    click_on "Delete this incomplete notification"
+    click_on "task list page"
+    click_on "Delete this draft"
     expect(page).to have_h1("Are you sure you want to delete #{draft_notification.product_name}?")
     click_button "Delete notification"
